requirements.txt
psycopg2
django>=1.11,<1.12
djangorestframework>=3.11,<3.12
django-extensions
django-mama-cas
django-taggit
django-bootstrap3
django-queryset-csv
django-admin-list-filter-dropdown
django_celery_beat
ckanapi
requests
timeout-decorator
pillow
lxml
redis
celery
-e git+https://github.com/neogeo-technologies/OWSLib.git@idgo#egg=OWSLib
markdown
xmltodict
# A supprimer quand passage Python 3.6+
python2-secrets
